backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
requests==2.31.0
python-dotenv==1.0.1
openai==1.30.1
langchain==0.1.17
langchain-community==0.0.34
langchain-openai==0.1.6
serpapi==0.3.4
